Create interfaces for Rank and Ranks, make them more generic

+++ TeamGenerator.Core/Interfaces/IEvaluate.cs
@@ -4,7 +4,7 @@ namespace TeamGenerator.Core.Interfaces
 {
     public interface IEvaluate
     {
-        int EvaluatePlayer(Player player);
-        int EvaluateTeam(Team team);
+        double EvaluatePlayer(Player player);
+        double EvaluateTeam(Team team);
     }
 }
